"Edit/delete bug fixed"

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,33 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
 
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        String error = "Only the owner of the image can edit the image";
+        Boolean Validate = Validation(image.getUser(), session);
+        if (!Validate) {
+            model.addAttribute("editError", error);
+            model.addAttribute("tags", tags);
+            return "images/image";
+        } else {
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+    }
+
+    private Boolean Validation(User user, HttpSession session) {
+
+        User sessionUser = (User) session.getAttribute("loggeduser");
+
+        if (user.getId() == sessionUser.getId()) {
+            return true;
+        } else
+            return false;
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +160,18 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        String deleteError = "Only the owner of the image can delete the image";
+        Image image = imageService.getImage(imageId);
+        Boolean isUser = Validation(image.getUser(), session);
+        if (!isUser) {
+            model.addAttribute("deleteError", deleteError);
+
+            return "images/image";
+        } else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
